+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Release date: 2019-03-XX.
 - [Keystore in withdraw command](https://github.com/eidoo/hybrid-exchange-sdk/issues/67) Update `withdraw` command using keystore instead of private key
 - [Keystore in deposit token command](https://github.com/eidoo/hybrid-exchange-sdk/issues/66) Update `deposit-token` command using keystore instead of private key
 - [Keystore in create wallet command](https://github.com/eidoo/hybrid-exchange-sdk/issues/65) Update `create-wallet` command using keystore instead of private key
+- [Keystore in order sign command](https://github.com/eidoo/hybrid-exchange-sdk/issues/63) Update `sign` command using keystore instead of private key
 - [CLI get allowance command](https://github.com/eidoo/hybrid-exchange-sdk/issues/59) Remove `privateKeyValidator` and `privateKeyService` dependencies from `GetAllowanceCommand`
 - [CLI get balance command](https://github.com/eidoo/hybrid-exchange-sdk/issues/58) Remove `privateKeyValidator` and `privateKeyService` dependencies from `GetBalanceCommand`
 - [CLI get address command](https://github.com/eidoo/hybrid-exchange-sdk/issues/47) Remove `privateKeyValidator` and `privateKeyService` dependencies from `GetAddressCommand`